requirements.txt
h5py==3.1.0
imbalanced-learn==0.7.0
imblearn==0.0
jams==0.3.4
librosa==0.8.0
matplotlib==3.3.3
numpy==1.19.2
opencv-python==4.5.1.48
pandas==1.1.3
PeakUtils==1.3.3
Pillow==8.1.0
PyYAML==5.3.1
scaper==1.6.4
scikit-learn==0.23.2
scipy==1.5.2
torch==1.7.1
torchaudio==0.7.2
tqdm==4.55.1
hydra-core==1.0.6
adjustText==0.7.3
